+++ core/src/main/scala/org/http4s/headers/ForwardedRenderers.scala
@@ -17,11 +17,10 @@
 package org.http4s.headers
 
 import java.nio.charset.StandardCharsets
-
 import cats.Eval
 import cats.syntax.flatMap._
 import org.http4s.Uri
-import org.http4s.parser.Rfc2616BasicRules
+import org.http4s.internal.parsing.Rfc7230
 import org.http4s.util.{Renderer, Writer}
 
 /** Renderers for the [[Forwarded]] header models.
@@ -80,7 +79,7 @@ private[http4s] trait ForwardedRenderers {
           writer << name << '='
 
           val rendered = Renderer.renderString(value)
-          if (Rfc2616BasicRules.isToken(rendered))
+          if (Rfc7230.token.parseAll(rendered).isRight)
             writer << rendered
           else
             writer <<# rendered // quote non-token values