+++ tests/src/test/scala/org/http4s/MediaTypeSpec.scala
@@ -56,22 +56,13 @@ class MediaTypeSpec extends Http4sSpec {
     }
 
     "parse literals" in {
-      val mediaType = MediaType.mediaType("application/json")
       val mediaTypeSC = mediaType"application/json"
-
-      mediaType must_== MediaType.application.`json`
       mediaTypeSC must_== MediaType.application.`json`
     }
 
     "reject invalid literals" in {
       import org.specs2.execute._, Typecheck._
       import org.specs2.matcher.TypecheckMatchers._
-
-      typecheck {
-        """
-           MediaType.mediaType("not valid")
-        """
-      } must not succeed
 
       typecheck {
         """